Rename individual search result
`Book` is really a variation of `SearchResult`, so I renamed it
`BookSearchResult`.  I did the same for `Review` and `User`.  I want to
introduce React Router so I can click on a search result to open a page
with more details, at which point I will have a `Book` widget that
represents a book and not just a search result.

I also switched the names of CSS classes from snake_case to kebab-case.